requirements.txt
flask
flask-cors
anthropic
pypdf
sentence-transformers
faiss-cpu
python-dotenv
langchain
langchain-anthropic
numpy